feat(debug): Add console logs for plan name display debugging
This commit introduces temporary console logs to assist in debugging why plan names are not being displayed in the DataTable on the \dmin/hotel-plans\ page.

Logs have been added to:
*   \rontend/src/pages/Admin/ManagePlans/ManagePlans.vue\:
    *   To inspect \plans.value\ after fetching.
    *   To inspect \hotelPlans.value\ after mapping.
    *   To inspect \slotProps.data\ within the \<Column field=\
name\>\ template of the DataTable.
*   \rontend/src/composables/usePlansStore.js\:
    *   To inspect the raw data returned from the API in \etchPlansForHotel\.

These logs will help in tracing the data flow and identifying where the plan names might be missing or incorrectly processed. This commit is intended for debugging purposes and the logs should be removed once the issue is identified and resolved.